subtask(F-1-7-web-generated-posture-surface): Move web posture UI to generated contract surface
