+++ apps/web/src/app/lib/api-base-url.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_TANREN_API_BASE_URL = "http://127.0.0.1:8081";
+
+export const TANREN_API_BASE_URL =
+  process.env["NEXT_PUBLIC_API_URL"] ?? DEFAULT_TANREN_API_BASE_URL;
