+++ docs/architecture/subsystems/interfaces.md
@@ -100,6 +100,10 @@ The web UI consumes the public API through generated TypeScript clients. It
 does not use a separate backend-for-frontend layer, maintain duplicate API
 shapes, or store server state as a divergent client-side source of truth.
 
+Operational controls use dedicated routes/components instead of overloading the
+root landing surface. Deployment-posture controls live at `/deployment-posture`
+and read/write posture state through the generated posture contract types.
+
 The web UI must be responsive enough to support desktop and phone workflows
 from the same application.
 
@@ -372,6 +376,10 @@ integration, or delivery models.
 Public API contracts are generated from Rust contract types and route metadata.
 The generated OpenAPI document is the source for TypeScript web clients and
 may also support external client generation.
+
+The web deployment-posture surface (`/deployment-posture`) consumes generated
+TypeScript contract types from the posture contract module, so it does not
+redeclare local DTO copies for request/response payloads.
 
 MCP tool schemas are generated from or declared against the same command/query
 contract layer. CLI and TUI command implementations call the API contract
